reference_frontend/gg-flutter-app-final_code/gg-flutter-app-final_code/lib/backend/Gluco-guide-main/requirements.txt
fastapi
uvicorn
pandas
matplotlib
sqlalchemy
pydantic
googletrans==4.0.0-rc1
torch
scikit-learn
numpy
easyocr
python-multipart
sentence-transformers